requirements.txt
pandas==2.3.3
numpy==2.2.3
scikit-learn==1.7.2
mlflow==3.5.1
flask==3.1.0
gunicorn==23.0.0
python-dotenv==1.0.1
pyyaml==6.0.3
joblib==1.5.2
matplotlib==3.10.7
altair==5.5.0
streamlit==1.51.0
pytest==8.4.2
flake8==7.1.1